Move JSC-dependent function to src/bun.js/bindings/
Create uws_bindings.cpp in src/bun.js/bindings/ where JavaScriptCore
headers are properly configured for all platforms including Windows.

The previous location in src/deps/libuwsockets.cpp caused build
failures on Windows because JavaScriptCore headers (specifically
bmalloc/BPlatform.h) were not available in that directory's
include paths.
